move Brigade to dedicated file for clarity
